Backend/requirements.txt.txt
absl-py==2.1.0
addict==2.4.0
aioice==0.9.0
aiortc==1.5.0
altair==5.0.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.1.0
av==10.0.0
basicsr==1.4.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
decorator==5.1.1
dnspython==2.4.1
easyocr==1.7.0
editdistance==0.8.1
efficientnet==1.0.0
email_validator==2.2.0
essential-generators==1.0
exceptiongroup==1.2.1
facexlib==0.3.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.12.2
filterpy==1.4.5
fonttools==4.41.1
future==1.0.0
gfpgan==1.3.8
gitdb==4.0.10
GitPython==3.1.32
google-crc32c==1.5.0
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.4
ifaddr==0.2.0
imageio==2.31.1
imgaug==0.4.0
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
Keras-Applications==1.0.8
keras-ocr==0.9.3
kiwisolver==1.4.4
lazy_loader==0.3
llvmlite==0.43.0
lmdb==1.4.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numba==0.60.0
numpy==1.25.1
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
orjson==3.10.5
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
platformdirs==4.2.2
protobuf==4.23.4
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyclipper==1.3.0.post4
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.8.0
pyee==11.0.0
Pygments==2.15.1
pylibsrtp==0.8.0
pymongo @ file:///C:/b/abs_cb77hedu0e/croot/pymongo_1714079769816/work
Pympler==1.0.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
realesrgan==0.3.0
referencing==0.30.0
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
scikit-image==0.21.0
scikit-learn==1.5.0
scipy==1.11.1
seaborn==0.12.2
shapely==2.0.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
starlette==0.37.2
streamlit==1.36.0
streamlit-webrtc==0.45.1
sympy==1.12
tb-nightly==2.18.0a20240616
tenacity==8.2.2
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2023.7.18
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==4.3.1
ujson==5.10.0
ultralytics==8.0.143
urllib3==2.0.4
uvicorn==0.30.1
validators==0.20.0
watchdog==3.0.0
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
yapf==0.40.2
zipp==3.16.2